Add spec to cover issue #20 (zero FIXMEs shouldn't be re-analysed)
