+++ spec/spec_helper.rb
@@ -5,3 +5,5 @@ require "rspec/collection_matchers"
 
 require File.join(File.dirname(__FILE__), "..", "lib", "preload")
 LOG.level = Logger::WARN
+
+require File.join(File.dirname(__FILE__), "integration", "integration_spec_support")
